requirements.txt
Flask==2.0.1
google-generativeai==0.3.1
gunicorn==20.1.0 